feat: enhance contact form
